Refactor solution for 2020 d1 and add missing input.data file for 2020 d2

+++ 2020/2020-1/src/main/java/DayOne.java
@@ -16,7 +16,7 @@ public class DayOne {
         int[] expenseReport = Files.lines(path).mapToInt(Integer::parseInt).toArray();
 
 
-        long[] solution = new Puzzle(expenseReport, 2020).solve();
+        int[] solution = new Puzzle(expenseReport, 2020).solve();
         System.out.println(Arrays.toString(solution));
     }
 }
